requirements.txt
bs4
requests~=2.27.1
pandas~=1.3.5
python-dotenv~=0.19.2

APScheduler~=3.8.1
beautifulsoup4~=4.10.0
pytz~=2021.3